backend_session/requirements.txt
django
djangorestframework
faker
